Create backbone of the main header
Created the backbone of the main header as well as added a main image to the header. Texts and links needs to be replaced eventually.
Created two Flexbox container inside the main header, which is .front-header and .top-header.
Those front and main header share the same background and some of the fonts however, the second header contains a image.
And therefore, I separated it so that i would have a better controll over the image.